Move flatten_dict to high-level utils module
